packages/display/requirements.txt
spidev
lgpio
Pillow
netifaces
